since zmq messages are immediately consumed, add code to reuse old data from a few iterations ago until the phantom app sends a new update message. this also serves as a redundant disconnection detection. might be able to remove the disconnect detection code from long_mpc
