+++ selfdrive/phantom_receiver.py
@@ -1,18 +1,23 @@
-import sys
 import zmq
 from selfdrive.services import service_list
 import selfdrive.messaging as messaging
 
-def broadcast_data(p_status, p_speed, p_angle, p_time):
-  p_status = True if p_status == "true" else False
-  phantomData_sock = messaging.pub_sock(zmq.Context(), service_list['phantomData'].port)
-  data = messaging.new_message()
-  data.init('phantomData')
-  data.phantomData.status = p_status
-  data.phantomData.speed = p_speed
-  data.phantomData.angle = p_angle
-  data.phantomData.time = p_time
-  phantomData_sock.send(data.to_bytes())
+class PhantomReceiver():
+  def __init__(self):
+    self.phantomData_sock = None
 
-if __name__ == "__main__":
-  broadcast_data(str(sys.argv[1]), float(sys.argv[2]), float(sys.argv[3]), float(sys.argv[4]))
+  def broadcast_data(self, status, speed, angle, time):
+    status = True if status == "true" or status is True else False
+    data = messaging.new_message()
+    data.init('phantomData')
+    data.phantomData.status = status
+    data.phantomData.speed = speed
+    data.phantomData.angle = angle
+    data.phantomData.time = time
+    self.phantomData_sock.send(data.to_bytes())
+
+  def open_socket(self):
+    self.phantomData_sock = messaging.pub_sock(zmq.Context(), service_list['phantomData'].port)
+
+  def close_socket(self):
+    self.phantomData_sock.close()
